% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.trame-package.R
\docType{package}
\name{data.trame-package}
\alias{data.trame-package}
\title{A better 'data.frame' for 'SciViews::R'}
\description{
The 'data.trame' object is an hybrid between 'data.table', tibble' and
'data.frame'. It enhances the 'data.frame' with the speed of 'data.table' and
the nice features of 'tibble' (petty printing, stricter rules...).
}
\section{Important functions}{

\itemize{
\item \code{\link[=data.trame]{data.trame()}} to construct a data.trame object,
\item \code{\link[=as.data.trame]{as.data.trame()}} to coerce into a data.trame object.
\item \code{\link[=is.data.trame]{is.data.trame()}} to test for data.trame objects.
\item Methods for data.frame objects.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/data.trame}
  \item \url{https://www.sciviews.org/data.trame/}
  \item \url{https://sciviews.r-universe.dev/data.trame}
  \item Report bugs at \url{https://github.com/SciViews/data.trame/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
