% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.trame.R
\name{data.trame}
\alias{data.trame}
\alias{as.data.trame}
\alias{as.data.trame.default}
\alias{is.data.trame}
\title{Build, coerce and test for 'data.trame' objects}
\usage{
data.trame(
  ...,
  .key = NULL,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

as.data.trame(
  x,
  .key = NULL,
  .rows = NULL,
  .rownames = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  ...
)

\method{as.data.trame}{default}(
  x,
  .key = NULL,
  .rows = NULL,
  .rownames = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  ...
)

is.data.trame(x)
}
\arguments{
\item{...}{A set of name-value pairs that constitute the 'data.trame'.}

\item{.key}{A character vector with the name of the columns to use as key
(sorting the data)}

\item{.rows}{The number of rows in the final 'data.trame'. Useful to create a
0-column object, or for additional check.}

\item{.name_repair}{Treatment of problematic column names: could be
\code{"check_unique"} (default, no name repair but check they are unique),
\code{"unique"} (make sure names are unique), \code{"universal"} (make sure names are
unique and syntactically correct), \code{"minimal"} (no check or name repair
except for existence), or a function for custom name repair, e.g.,
\code{.name.repair = make.names} for base R style.}

\item{x}{An object to test.}

\item{.rownames}{The name of the column that holds the row names, if any.}
}
\value{
A 'data.trame' object, which is indeed a
'data.trame'/'data.table'/'data.frame' object, thus subclassing both
'data.table' and 'data.frame'. \code{is.data.trame()} returns \code{TRUE} if the
object is a 'data.trame', and \code{FALSE} otherwise.
}
\description{
Build, coerce and test for 'data.trame' objects
}
\examples{
dtrm <- data.trame(
  a = 1:3,
  b = letters[1:3],
  c = factor(LETTERS[1:3]), .key = c('a', 'b'), .rows = 3
)
is.data.trame(dtrm)
}
