% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcast.R
\name{dcast}
\alias{dcast}
\alias{dcast.data.trame}
\title{Fast dcast for data.trame}
\usage{
dcast(
  data,
  formula,
  fun.aggregate = NULL,
  ...,
  margins = NULL,
  subset = NULL,
  fill = NULL,
  value.var = guess(data)
)

\method{dcast}{data.trame}(
  data,
  formula,
  fun.aggregate = NULL,
  sep = "_",
  ...,
  margins = NULL,
  subset = NULL,
  fill = NULL,
  drop = TRUE,
  value.var = guess(data),
  verbose = getOption("datatable.verbose")
)
}
\arguments{
\item{data}{A data.trame object.}

\item{formula}{A formula LHS ~ RHS, see \code{\link[data.table:dcast.data.table]{data.table::dcast()}}
for details.}

\item{fun.aggregate}{Te function to aggregate multiple data before dcasting.}

\item{...}{Arguments passed to the aggregating function.}

\item{margins}{Note implemented yet.}

\item{subset}{Should the dcasting be done on a subset of the data?}

\item{fill}{Value with which to fill missing cells.}

\item{value.var}{The name of the column to use as value variable. If not
provided it is "guessed" the \code{guess()} internal function is used to build
a good default name.}

\item{sep}{Character vector of length 1, used to separate parts of the
variable names (\verb{_} by default).}

\item{drop}{\code{FALSE} should dcast include all missing combinations?}

\item{verbose}{Not used yet.}
}
\value{
A keyed data.trame is returned with the dcasted data.
}
\description{
\code{dcast()} transforms a data.trame from long to wide format, a little bit like
\code{tidyr::pivot_wider()}. See \code{\link[data.table:dcast.data.table]{data.table::dcast()}} for explanations.
}
\examples{
# Adapted from first example of ?dcast.data.table
ChickWeight = as.data.trame(ChickWeight)
ChickWeight <- set_names(ChickWeight, tolower(names(ChickWeight)))
dtrm <- melt(ChickWeight, id.vars = 2:4)
dcast(dtrm, time ~ variable, fun.aggregate = mean)
}
\seealso{
\code{\link[data.table:dcast.data.table]{data.table::dcast()}}, \code{\link[=melt]{melt()}}
}
