% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.trame-utils.R
\name{let_data.table_to_data.trame}
\alias{let_data.table_to_data.trame}
\alias{let_data.trame_to_data.table}
\title{Switch back and forth quickly between data.trame and data.table by reference}
\usage{
let_data.table_to_data.trame(x)

let_data.trame_to_data.table(x)
}
\arguments{
\item{x}{A valid data.trame or data.table object}
}
\value{
A data.table or data.trame object.
}
\description{
These functions just change the class of the object by reference (data.trame
and data.table objects are, internally, identical except for their class).
These functions are intended only for programmers!
}
\examples{
dtrm <- data.trame(a = 1:3, b = letters[1:3], c = factor(LETTERS[1:3]))
class(dtrm)
let_data.trame_to_data.table(dtrm)
class(dtrm)
let_data.table_to_data.trame(dtrm)
class(dtrm)

# Whenever you need to use a data.table method on a data.trame in a function,
# you can do something like this:
test_fun <- function(x, i, expr) {
  force(i) # Make sure i is evaluated before changing the class of x
  # value will be evaluated with x being a data.table here!
  let_data.trame_to_data.table(x)
  on.exit(let_data.table_to_data.trame(x))
  expr
}
test_fun(dtrm, 1:2, class(dtrm))
class(dtrm) # Back to data.trame
}
