% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.data.trame.R
\name{print.data.trame}
\alias{print.data.trame}
\alias{format.data.trame}
\alias{obj_sum.data.trame}
\alias{tbl_sum.data.trame}
\alias{tbl_nrow.data.trame}
\alias{str.data.trame}
\title{Printing data.trames}
\usage{
\method{print}{data.trame}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{format}{data.trame}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{obj_sum}{data.trame}(x)

\method{tbl_sum}{data.trame}(x, ...)

\method{tbl_nrow}{data.trame}(x, ...)

\method{str}{data.trame}(object, ..., indent.str = " ", nest.lev = 0)
}
\arguments{
\item{x}{A data.trame object.}

\item{width}{The width of the text output. If \code{NULL}, the default,
\code{getOption("width")} is used.}

\item{...}{Additional arguments passed to \code{format()}.}

\item{n}{The number of rows to print. If \code{NULL}, a default number is used.}

\item{max_extra_cols}{The maximum number of extra columns to print
abbreviated. If \code{NULL} by default, a reasonable default value is used.}

\item{max_footer_lines}{Maximum number of lines in the footer. If \code{NULL}, a
reasonable default value is used.}

\item{object}{A data.trame.}

\item{indent.str}{The string used for indentation.}

\item{nest.lev}{The current nesting level, used for recursive printing.}
}
\description{
A data.trame prints almost like a tibble.
}
\examples{
dtrm <- data.trame(
  a = -1:3,
  b = letters[1:5],
  c = factor(LETTERS[1:5]),
  d = c(TRUE, FALSE, TRUE, NA, TRUE), .key = c('a', 'b'), .rows = 5
)
dtrm
str(dtrm)
}
